requirements.txt
Django==2.0.1
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
docutils==0.14
jmespath==0.9.3
Markdown==2.6.11
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
